scraper/requirements.txt
-i https://pypi.org/simple
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15; python_version >= '3.6'
charset-normalizer==2.1.1; python_version >= '3.6'
filelock==3.8.0; python_full_version >= '3.7.0'
huggingface-hub==0.9.1; python_full_version >= '3.7.0'
idna==3.3; python_version >= '3.5'
numpy==1.23.2; python_version >= '3.8'
packaging==21.3; python_version >= '3.6'
pillow==9.2.0; python_full_version >= '3.7.0'
psycopg2==2.9.3
pyparsing==3.0.9; python_full_version >= '3.6.8'
python-dotenv==0.20.0
pyyaml==6.0; python_version >= '3.6'
regex==2022.8.17; python_version >= '3.6'
requests==2.28.1
soupsieve==2.3.2.post1; python_version >= '3.6'
sqlalchemy==1.4.40
tokenizers==0.12.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
transformers==4.21.2
typing-extensions==4.3.0; python_full_version >= '3.7.0'
urllib3==1.26.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
